ci: fix failures on Dependabot PRs
Dependabot PRs are treated as forks and don't have access to repo
secrets, including the secret required to configure AWS access and do
the Lambda deploy dry-run.

Since the deploy action does not involve the NPM dependencies of the
project, there should be no need to run it on Dependabot PRs anyway, so
we can disable it in that context.

Also fix a mistake where the deploy dry-run would not even be attempted
if the `dry-run` option was set, defeating the point.