ci: fix deploy collisions
On pushes to the main branch, the dry-run deploy in the CI workflow and
the real deploy from the Deploy workflow could run at nearly the same
time, resulting in one of them throwing a `ResourceConflictException`.
We don't need to do the dry run when we are already doing a "real" run
as part of the same push.